LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/pixel-manipulation/2d.imageData.create2.zero.worker-expected.txt


PASS createImageData(sw, sh) throws INDEX_SIZE_ERR if size is zero
